ipq807x: add PCI support
This is a heavy mix of backports and stuff pending
for IPQ6018 which uses the same IP for Gen3 PCIe.

Gen2 PCIe has upstream support, but it requires fixes.
Namely, it requires PHY to be started as well as resets and clocks to be configured before trying to
access the PCI register space.

Signed-off-by: Robert Marko <robimarko@gmail.com>